+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -17,7 +17,7 @@
 ### Other Changes
 
 - Create separate lists of characters that are allowed within tenant ids and scopes in `AzureCliCredential`.
-- Add default values to some `WorkloadIdentityCredentialOption` fields such as authority host by reading them from the environment.
+- Add default values to some `WorkloadIdentityCredentialOptions` fields such as authority host by reading them from the environment.
 - Add logging to `WorkloadIdentityCredential` to help with debugging.
 
 ## 1.6.0-beta.3 (2023-10-12)